componentize core utility functions
