--- src/utils/core.js
@@ -1,81 +1,0 @@
-export function isElementVisible (element) {
-	const container = this.store.containers[element.containerId]
-	const viewFactor = Math.max(0, Math.min(1, element.config.viewFactor))
-	const viewOffset = element.config.viewOffset
-
-	const elementBounds = {
-		top: element.geometry.bounds.top + element.geometry.height * viewFactor,
-		right: element.geometry.bounds.right - element.geometry.width * viewFactor,
-		bottom: element.geometry.bounds.bottom - element.geometry.height * viewFactor,
-		left: element.geometry.bounds.left + element.geometry.width * viewFactor,
-	}
-
-	const containerBounds = {
-		top: container.geometry.bounds.top + container.scroll.top + viewOffset.top,
-		right: container.geometry.bounds.right + container.scroll.left - viewOffset.right,
-		bottom: container.geometry.bounds.bottom + container.scroll.top - viewOffset.bottom,
-		left: container.geometry.bounds.left + container.scroll.left + viewOffset.left,
-	}
-
-	return (
-		(elementBounds.top < containerBounds.bottom &&
-			elementBounds.right > containerBounds.left &&
-			elementBounds.bottom > containerBounds.top &&
-			elementBounds.left < containerBounds.right) ||
-		element.styles.position === 'fixed'
-	)
-}
-
-export function getGeometry (target, isContainer) {
-	/**
-	 * We want to ignore padding and scrollbars for container elements.
-	 * More information here: https://goo.gl/vOZpbz
-	 */
-	const height = isContainer ? target.node.clientHeight : target.node.offsetHeight
-	const width = isContainer ? target.node.clientWidth : target.node.offsetWidth
-
-	let offsetTop = 0
-	let offsetLeft = 0
-	let node = target.node
-
-	do {
-		if (!isNaN(node.offsetTop)) {
-			offsetTop += node.offsetTop
-		}
-		if (!isNaN(node.offsetLeft)) {
-			offsetLeft += node.offsetLeft
-		}
-		node = node.offsetParent
-	} while (node)
-
-	return {
-		bounds: {
-			top: offsetTop,
-			right: offsetLeft + width,
-			bottom: offsetTop + height,
-			left: offsetLeft,
-		},
-		height,
-		width,
-	}
-}
-
-export function getScrolled (container) {
-	return container.node === document.documentElement
-		? {
-			top: window.pageYOffset,
-			left: window.pageXOffset,
-		}
-		: {
-			top: container.node.scrollTop,
-			left: container.node.scrollLeft,
-		}
-}
-
-export function logger (message, ...details) {
-	if (this.constructor.debug && console) {
-		let report = `%cScrollReveal: ${message}`
-		details.forEach(detail => (report += `\n — ${detail}`))
-		console.log(report, 'color: #ea654b;') // eslint-disable-line no-console
-	}
-}
