+++ src/utils/get-geometry.js
@@ -1,0 +1,33 @@
+export default function getGeometry (target, isContainer) {
+	/**
+	 * We want to ignore padding and scrollbars for container elements.
+	 * More information here: https://goo.gl/vOZpbz
+	 */
+	const height = isContainer ? target.node.clientHeight : target.node.offsetHeight
+	const width = isContainer ? target.node.clientWidth : target.node.offsetWidth
+
+	let offsetTop = 0
+	let offsetLeft = 0
+	let node = target.node
+
+	do {
+		if (!isNaN(node.offsetTop)) {
+			offsetTop += node.offsetTop
+		}
+		if (!isNaN(node.offsetLeft)) {
+			offsetLeft += node.offsetLeft
+		}
+		node = node.offsetParent
+	} while (node)
+
+	return {
+		bounds: {
+			top: offsetTop,
+			right: offsetLeft + width,
+			bottom: offsetTop + height,
+			left: offsetLeft,
+		},
+		height,
+		width,
+	}
+}
